+++ confluence/charts/src/__tests__/integration/index.ts
@@ -1,0 +1,21 @@
+import { BrowserTestCase } from '@atlaskit/webdriver-runner/runner';
+import { getExampleUrl } from '@atlaskit/webdriver-runner/utils/example';
+import Page from '@atlaskit/webdriver-runner/wd-wrapper';
+
+/* Url to test the example */
+const testUrl = getExampleUrl('confluence', 'charts', 'basic');
+
+/* Css selectors used for the test */
+const exampleComponent = "[data-testid='charts']";
+
+BrowserTestCase(
+  'Charts should be able to be identified by data-testid',
+  {},
+  async (client: any) => {
+    const componentTest = new Page(client);
+    await componentTest.goto(testUrl);
+
+    // Check for visibility.
+    expect(await componentTest.isVisible(exampleComponent)).toBe(true);
+  },
+);